+++ lib/src/commands/make/plugin.dart
@@ -35,7 +35,7 @@ class PluginCommand extends Command {
     }
 
     List<MakerDependency> deps = [
-      const MakerDependency('angel_framework', '^1.0.0')
+      const MakerDependency('angel_framework', '^2.0.0')
     ];
 
     var rc = new ReCase(name);