+++ lib/src/commands/service_generators/file_service.dart
@@ -8,7 +8,7 @@ class FileServiceGenerator extends ServiceGenerator {
 
   @override
   List<MakerDependency> get dependencies =>
-      const [const MakerDependency('angel_file_service', '^1.0.0')];
+      const [const MakerDependency('angel_file_service', '^2.0.0')];
 
   @override
   bool get goesFirst => true;